leo_python/requirements.txt
selenium
pandas
PyDrive
python-sonarqube-api